Move from ES module -> CommonJS
- Move from ES module to CommonJS

Below is the detailed rational:

- in ESM, you need the full name to import something. For example
  `import "src/routes/get.js"`. Since, we are working in typescript, it
creates confusion. The "src/routes/get.js" file is created by `tsc` on
compilation and until then it has no existance. Therefore, it is
confusing.

- we can still use the `import syntax` since typescript will
  automatically convert ESM imports to `require` for us.

- we do not have to include the keyword `type` when importing types

These changes come with ramifications that have far reaching
consequences for the project. So we might quickly iterate and may come
back to ESM later.